+++ packages/selected-network-controller/tests/SelectedNetworkController.test.ts
@@ -527,33 +527,46 @@ describe('SelectedNetworkController', () => {
       });
 
       describe('when the requesting domain is a snap (starts with "npm:" or "local:"', () => {
-        it('skips setting the networkClientId for the passed in domain', () => {
+        it('sets the networkClientId for the passed in snap ID', () => {
           const { controller, mockHasPermissions } = setup({
             state: { domains: {} },
             useRequestQueuePreference: true,
           });
           mockHasPermissions.mockReturnValue(true);
-          const snapDomainOne = 'npm:@metamask/bip32-example-snap';
-          const snapDomainTwo = 'local:@metamask/bip32-example-snap';
-          const nonSnapDomain = 'example.com';
+          const domain = 'npm:foo-snap';
           const networkClientId = 'network1';
+          controller.setNetworkClientIdForDomain(domain, networkClientId);
+          expect(controller.state.domains[domain]).toBe(networkClientId);
+        });
 
-          controller.setNetworkClientIdForDomain(
-            nonSnapDomain,
-            networkClientId,
-          );
-          controller.setNetworkClientIdForDomain(
-            snapDomainOne,
-            networkClientId,
-          );
-          controller.setNetworkClientIdForDomain(
-            snapDomainTwo,
-            networkClientId,
-          );
-
-          expect(controller.state.domains).toStrictEqual({
-            [nonSnapDomain]: networkClientId,
+        it('updates the provider and block tracker proxy when they already exist for the snap ID', () => {
+          const { controller, mockProviderProxy, mockHasPermissions } = setup({
+            state: { domains: {} },
+            useRequestQueuePreference: true,
           });
+          mockHasPermissions.mockReturnValue(true);
+          const initialNetworkClientId = '123';
+
+          // creates the proxy for the new domain
+          controller.setNetworkClientIdForDomain(
+            'npm:foo-snap',
+            initialNetworkClientId,
+          );
+          const newNetworkClientId = 'abc';
+
+          expect(mockProviderProxy.setTarget).toHaveBeenCalledTimes(1);
+
+          // calls setTarget on the proxy
+          controller.setNetworkClientIdForDomain(
+            'npm:foo-snap',
+            newNetworkClientId,
+          );
+
+          expect(mockProviderProxy.setTarget).toHaveBeenNthCalledWith(
+            2,
+            expect.objectContaining({ request: expect.any(Function) }),
+          );
+          expect(mockProviderProxy.setTarget).toHaveBeenCalledTimes(2);
         });
       });
 
@@ -776,23 +789,22 @@ describe('SelectedNetworkController', () => {
 
     // TODO - improve these tests by using a full NetworkController and doing more robust behavioral testing
     describe('when the domain is a snap (starts with "npm:" or "local:")', () => {
-      it('returns a proxied globally selected networkClient and does not create a new proxy in the domainProxyMap', () => {
-        const { controller, domainProxyMap, messenger } = setup({
+      it('calls to NetworkController:getSelectedNetworkClient and creates a new proxy provider and block tracker with the proxied globally selected network client', () => {
+        const { controller, messenger } = setup({
           state: {
             domains: {},
           },
-          useRequestQueuePreference: true,
+          useRequestQueuePreference: false,
         });
         jest.spyOn(messenger, 'call');
-        const snapDomain = 'npm:@metamask/bip32-example-snap';
 
-        const result = controller.getProviderAndBlockTracker(snapDomain);
-
-        expect(domainProxyMap.get(snapDomain)).toBeUndefined();
+        const result = controller.getProviderAndBlockTracker('npm:foo-snap');
+        expect(result).toBeDefined();
+        // unfortunately checking which networkController method is called is the best
+        // proxy (no pun intended) for checking that the correct instance of the networkClient is used
         expect(messenger.call).toHaveBeenCalledWith(
           'NetworkController:getSelectedNetworkClient',
         );
-        expect(result).toBeDefined();
       });
 
       it('throws an error if the globally selected network client is not initialized', () => {